+++ crates/core/src/integrators.rs
@@ -2,8 +2,8 @@
 
 use nalgebra::Vector3;
 
-use crate::potential::Potentials;
-use crate::property::{Forces, Property};
+use crate::potentials::Potentials;
+use crate::properties::{Forces, Property};
 use crate::system::System;
 
 /// A numerical integration algorithm.
@@ -15,8 +15,6 @@ pub trait Integrator {
 }
 
 /// Velocity Verlet integration algorithm.
-///
-/// Include equations here.
 #[derive(Clone, Debug)]
 pub struct VelocityVerlet {
     timestep: f32,
@@ -24,7 +22,11 @@ pub struct VelocityVerlet {
 }
 
 impl VelocityVerlet {
-    /// Returns a new `VelocityVerlet`.
+    /// Returns a new velocity verlet algorithm.
+    ///
+    /// # Arguments
+    ///
+    /// * `timestep` - Timestep duration
     pub fn new(timestep: f32) -> VelocityVerlet {
         VelocityVerlet {
             timestep,
@@ -49,12 +51,11 @@ impl Integrator for VelocityVerlet {
         }
 
         // calculate forces
-        let forces = Forces;
-        let forces = forces.calculate(system, potentials);
+        let forces = Forces.calculate(system, potentials);
 
         // update accelerations at t + dt
         for i in 0..sys_size {
-            self.accelerations[i] = forces[i] / system.masses[i];
+            self.accelerations[i] = forces[i] / system.elements[i].mass();
         }
 
         // update velocities at t + dt
@@ -66,52 +67,16 @@ impl Integrator for VelocityVerlet {
 
 #[cfg(test)]
 mod tests {
-    use crate::integrate::{Integrator, VelocityVerlet};
-    use crate::potential::pair::{LennardJones, PairPotentialMeta};
-    use crate::potential::{Potentials, Restriction};
-    use crate::system::{cell::Cell, element::Element, System};
-    use nalgebra::Vector3;
-
-    fn get_pair_system() -> System {
-        let size = 2 as usize;
-        let argon = Element::Ar;
-        let mut sys = System::new(size);
-        sys.cell = Cell::new(17.0, 17.0, 17.0, 90.0, 90.0, 90.0);
-        sys.elements = vec![argon, argon];
-        sys.molecules = vec![0 as usize, 0 as usize];
-        sys.positions = vec![Vector3::new(0.0, 0.0, 0.0), Vector3::new(3.4, 3.4, 3.4)];
-        sys.velocities = vec![
-            Vector3::new(
-                -0.007225222699367925,
-                -0.002405756495275919,
-                0.0026065109398392215,
-            ),
-            Vector3::new(
-                0.001179633958023287,
-                0.003525262341736351,
-                -0.0004132774783154952,
-            ),
-        ];
-        sys.masses = vec![argon.mass(), argon.mass()];
-        sys.charges = vec![0.0, 0.0];
-        sys
-    }
-
-    fn get_pair_potentials() -> Potentials {
-        let mut pots = Potentials::new();
-        let potential = Box::new(LennardJones::new(1.0, 3.4));
-        let meta = PairPotentialMeta::new((Element::Ar, Element::Ar), 8.5, Restriction::None);
-        pots.add_pair(potential, meta);
-        pots
-    }
+    use super::{Integrator, VelocityVerlet};
+    use crate::utils::{load_test_potentials, load_test_system};
 
     #[test]
     fn velocity_verlet() {
         // define the system
-        let mut sys = get_pair_system();
+        let mut sys = load_test_system("argon");
 
         // define the potentials
-        let pots = get_pair_potentials();
+        let pots = load_test_potentials("argon");
 
         // define the integrator
         let mut vv = VelocityVerlet::new(1.0);