+++ crates/core/src/lib.rs
@@ -1,7 +1,12 @@
 //! Core feature library for the Velvet simulation engine.
 
-#![deny(missing_docs)]
-pub mod integrate;
-pub mod potential;
-pub mod property;
+#[warn(missing_docs)]
+pub mod constants;
+pub mod distributions;
+pub mod integrators;
+pub mod potentials;
+pub mod properties;
 pub mod system;
+pub mod thermostats;
+
+pub mod utils;