+++ crates/core/src/potentials/mod.rs
@@ -1,13 +1,17 @@
-//! Interatomic potentials for evaluating potential energy and forces.
+//! Interatomic potentials to evaluate potential energy and forces.
 
 pub mod pair;
 
-use crate::potential::pair::{PairPotential, PairPotentialMeta};
+use serde::{Deserialize, Serialize};
+
+use crate::potentials::pair::{PairPotential, PairPotentialMeta};
 
 /// Base trait for all potentials.
+#[typetag::serde(tag = "type")]
 pub trait Potential: Send + Sync {}
 
 /// Container type to hold instances of each potential in the system.
+#[derive(Serialize, Deserialize)]
 pub struct Potentials {
     pairs: Vec<(Box<dyn PairPotential>, PairPotentialMeta)>,
     // --bond--
@@ -17,7 +21,7 @@ pub struct Potentials {
 }
 
 impl Potentials {
-    /// Returns a new `Potentials`.
+    /// Returns an empty collection of potentials.
     pub fn new() -> Potentials {
         Potentials { pairs: Vec::new() }
     }
@@ -27,14 +31,19 @@ impl Potentials {
         self.pairs.iter()
     }
 
-    /// Adds a new pair potentials to the collection.
+    /// Adds a new pair potential to the collection.
+    ///
+    /// # Arguments
+    ///
+    /// * `potential` - Boxed pair potential trait object
+    /// * `meta` - Pair potential metadata
     pub fn add_pair(&mut self, potential: Box<dyn PairPotential>, meta: PairPotentialMeta) {
         self.pairs.push((potential, meta))
     }
 }
 
-/// Restrictions which can be applied to any potential.
-#[derive(Clone, Copy, Debug)]
+/// Restrictions on the applicability of a potential.
+#[derive(Clone, Copy, Debug, Serialize, Deserialize)]
 pub enum Restriction {
     /// No restrictions.
     None,