+++ crates/core/src/system/elements.rs
@@ -1,7 +1,10 @@
 //! Elemental properties.
 
+use serde::{Deserialize, Serialize};
+use strum_macros::EnumString;
+
 /// Every element on the periodic table.
-#[derive(Clone, Copy, Debug, PartialEq)]
+#[derive(Clone, Copy, Debug, PartialEq, Serialize, Deserialize, EnumString)]
 pub enum Element {
     /// Hydrogen
     H,
@@ -22,5 +25,17 @@ impl Element {
             Element::F => 18.998,
             Element::Ar => 39.948,
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::Element;
+    use std::str::FromStr;
+
+    #[test]
+    fn from_str() {
+        let hydrogen = Element::from_str("H").unwrap();
+        assert_eq!(Element::H, hydrogen)
     }
 }