+++ examples/nvt.rs
@@ -1,0 +1,115 @@
+//! This example executes a simulation of 108 Ar atoms in the NVT ensemble.
+//! After sucessful completion, a figure is generated at `nvt.png` which plots the total energy at each timestep.
+
+extern crate pretty_env_logger;
+#[macro_use]
+extern crate log;
+use indicatif::{ProgressBar, ProgressStyle};
+use plotters::prelude::*;
+
+use velvet_convert::load_poscar;
+use velvet_core::distributions::{Boltzmann, VelocityDistribution};
+use velvet_core::integrators::{Integrator, VelocityVerlet};
+use velvet_core::potentials::pair::{LennardJones, PairPotentialMeta};
+use velvet_core::potentials::{Potentials, Restriction};
+use velvet_core::properties::{Property, Temperature};
+use velvet_core::system::elements::Element;
+use velvet_core::thermostats::{Thermostat, NoseHoover};
+
+static TIMESTEPS: u64 = 250000;
+static PLOT_INTERVAL: u64 = 10;
+static FILENAME: &'static str = "assets/nvt.png";
+
+fn main() {
+    pretty_env_logger::init();
+    info!("Starting a NVT simulation of Ar gas...");
+
+    // Load the Ar gas system directly from a POSCAR formatted file.
+    let mut system = load_poscar("resources/test/argon.poscar");
+
+    // Setup an initial velocity distribution with a target temperature
+    let boltz = Boltzmann::new(300 as f32);
+
+    // Apply the initialized velocity distribution to the system.
+    boltz.apply(&mut system);
+
+    // Define a Lennard-Jones style pair potential.
+    let lj = LennardJones::new(1.0, 3.4);
+
+    // Define some metadata about the potential.
+    // - The element pair which it applies to.
+    // - The cutoff radius.
+    // - Any additional restrictions (intermolecular/intramolecular...)
+    let meta = PairPotentialMeta::new((Element::Ar, Element::Ar), 8.5, Restriction::None);
+
+    // Initialize a collection of potentials and add the previously defined pair potential with metadata.
+    let mut potentials = Potentials::new();
+    potentials.add_pair(Box::new(lj), meta);
+
+    // Define a velocity Verlet style integrator.
+    let mut velocity_verlet = VelocityVerlet::new(1.0);
+    velocity_verlet.setup(&system, &potentials);
+
+    // Define a Nose-Hoover thermostat
+    let mut nose_hoover = NoseHoover::new(300 as f32, 1.5, 1.0);
+    nose_hoover.setup(&system);
+
+    // Setup a progress bar to track the simulation.
+    let progress = get_progress_bar(TIMESTEPS);
+
+    let mut temperature_results: Vec<(u64, f64)> =
+        Vec::with_capacity((TIMESTEPS / PLOT_INTERVAL) as usize);
+
+    // Integrate for N timesteps.
+    for i in 0..TIMESTEPS {
+        nose_hoover.pre_integrate(&mut system);
+        velocity_verlet.integrate(&mut system, &potentials);
+        nose_hoover.post_integrate(&mut system);
+        if i % PLOT_INTERVAL == 0 {
+            temperature_results.push((i, Temperature.calculate(&system, &potentials) as f64));
+        }
+        progress.inc(1);
+    }
+
+    progress.finish();
+
+    info!("Simulation completed successfully.");
+
+    // Plot the energy results
+    plot_results(temperature_results);
+
+    info!("Generated summary figure: `{}`", FILENAME);
+}
+
+fn get_progress_bar(len: u64) -> ProgressBar {
+    let progress = ProgressBar::new(len);
+    progress.set_style(
+        ProgressStyle::default_bar()
+            .template("[{elapsed_precise}] {bar:40.green} {pos:>6}/{len} timesteps"),
+    );
+    progress
+}
+
+fn plot_results(data: Vec<(u64, f64)>) {
+    let root_area = BitMapBackend::new(FILENAME, (1024, 768)).into_drawing_area();
+    root_area.fill(&WHITE).unwrap();
+
+    let mut chart = ChartBuilder::on(&root_area)
+        .set_label_area_size(LabelAreaPosition::Left, 100)
+        .set_label_area_size(LabelAreaPosition::Bottom, 50)
+        .margin(10)
+        .margin_right(30)
+        .build_cartesian_2d(0..TIMESTEPS, 275.0..325.0)
+        .unwrap();
+
+    chart
+        .configure_mesh()
+        .x_desc("Timestep")
+        .x_label_style(("sans-serif", 18))
+        .y_desc("Temperature (K)")
+        .y_label_style(("sans-serif", 18))
+        .draw()
+        .unwrap();
+
+    chart.draw_series(LineSeries::new(data.into_iter(), &BLUE)).unwrap();
+}